requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
matplotlib>=3.4.0
pillow>=8.0.0
tqdm>=4.60.0
pandas>=1.3.0
seaborn>=0.11.0
colorthief>=0.2.1
webcolors>=1.11.0 